+++ examples/custom-toolchain/flake.nix
@@ -65,7 +65,7 @@
           '';
         };
 
-        devShell = pkgs.mkShell {
+        devShells.default = pkgs.mkShell {
           inputsFrom = builtins.attrValues self.checks;
 
           # Extra inputs can be added here